GCP/cloudRuns/generate_pdf/requirements.txt
Flask==2.2.5
fpdf==1.7.2
google-cloud-storage==2.11.0